Improved the application on mobile devices
When hoovering above a name it shows when the user arrived and left. But
on majority on mobile phones there is no hoover function and it is then
not possible to get this information. The change i made now shows the
arrival and leaving after the name on mobile devices.